requirements.txt
certifi==2021.10.8
click==8.1.2
colorama==0.4.4
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
numpy==1.21.6
pandas==1.3.5
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.35
typing_extensions==4.2.0
Werkzeug==2.1.1
wincertstore==0.2
zipp==3.8.0
